# openapi.yaml
openapi: 3.1.0
info:
  title: CraftifAI Control Plane API
  version: 0.1.0
  description: |
    Multi-tenant AI control plane. All authenticated routes require a session cookie
    (`craftifai_session`). Multi-org users may send `X-Org-Id` to select context.
    Cross-tenant access returns 404 (not 403).
servers:
  - url: http://localhost/api
    description: Docker Compose (nginx proxy)
  - url: http://localhost:3000
    description: Local API dev server
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register user and new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                display_name: { type: string, minLength: 1 }
      responses:
        '201':
          description: Created; session cookie set
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  org_id: { type: string, format: uuid }
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: OK; session cookie set
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      tags: [auth]
      security: [{ cookieAuth: [] }]
      responses:
        '204':
          description: Session revoked
  /auth/me:
    get:
      tags: [auth]
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: craftifai_session
...
  schemas:
    Session:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        org_id: { type: string, format: uuid }
        role: { type: string, enum: [administrator, member] }
...